Criando funcionalidade de buscar medida por ID

+++ back_end/cokenary/src/main/java/com/receitas/repository/MedidaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface MedidaRepository extends JpaRepository<Medida, Integer> {
+public interface MedidaRepository extends JpaRepository<Medida, Long> {
 
 }
